Refactored Constructor and Functional Interface
Refactored the constructor and the functional interface in preparation
for the addition of presets. The dictionary path has been removed from
the config, and a getter/setter for the dictionary file has been added.
The architecture for presets is all in place, but only one preset
exists so far, 'DEFAULT'.