Presets Tweaked to generate sufficient entropy
All presets have been tweaked to generate sufficient entropy to avoid
warnings with the exception of NTLM which can never possibly generate
enough entropy in the case where the attacker knows the dictionary and
preset because of it's 14 character limit.

In the process of tweaking these presets also discovered (and patched)
a bug in the entropy calculation function that was significantly
under-stating the entropy in the case where the attacker has the
dictionary and config.