Added padding_alphabet config key
To make the configs more readable, there are now three alphabet keys -
the original symbol_alphabet still exists, and is used as the fallback
if either padding_character is set to RANDOM but padding_alphabet is
not set, or separator_character is set to RANDOM but separator_alphabet
is not set.

Practically speaking, if you want to use the same alphabet for both
random separators and random padding characters, use symbol_alphabet,
if you want separate alphabets for each, use padding_alphabet and
separator_alphabet.