Continued to implement Type::Params & Tweaked PerlCritic Config
All public functions in Crypt::HSXKPasswd (with the exception of the
constructor) have now been ported over to use Type::Params for
parameter validation.

I tweaked the PerlCritic config for the project to make it a little
harsher in some areas, and then tweaked some code as needed to remove
errors.