Continued move to Type::Params, added depublication to config and dictionary setters
With the exception of the constructor, all subs in Crypt::HSXKPasswd
now use Type::Params for validation, including the private ones.

A new function has been added to the helper to make detection of debug
mode easy.

Features have been added to all setters for dictionaries and configs to
deduplicate words and symbol alphabets. To facilitate this, new
functions to distil word lists and symbol alphabets were added and
documented.