Added hsxkpasswdrc files, fixed entropy calc bug
Added support for hsxkpasswdrc files to allow custom presets and custom
defaults be set for the terminal command.

While making this change, also changed the way entropy warnings are
controlled. The old way was inverted - it was what you wanted to
suppress, not what you wanted to emit. This was just too confusing, so
have inverted the logic so the module config variable now controls what
should be emitted.

Added a little more entropy to the ALTERNATE case transform by
randomising the case of the first word, and then alternating from that
starting point.

While updating the entropy calculations for the ALTERNATE case
transform, found and fixed a bug in the entropy calculation that was
causing it to over-estimate the SEEN entropy.

Fixing this entropy calculation bug required some of the presets to be
tweaked to get them back above the warning thresholds. Unfortunately,
with the current default dictionary, WEB16 can no longer get under the
warning threshold, so it is now considered less secure, just like the
NTLM preset.