Improved CLI & hsxkpasswdrc files
Added a --dict-file option to allow a dictionary file be specified from
the command line.

Changed the --dictionary-package and --rng-package options to
--dict-pkg and --rng-pkg. Also changed -d to it is now an alias to
--dict-file.

Added --dict-pkg-args and --rng-pkg-arg options so arguments can now be
specified for the constructors for the dictionary and RNG packages.

Added support for all the above new options into hsxkpasswdrc files.