Remove Makefile.PL entry from MANIFEST
This project uses `Module::Build` which has no `Makefile.PL` file, hence
the entry in the MANIFEST was unnecessary.  Also, because of this extra
entry in the MANIFEST, the build script produced this warning:

```
$ perl Build.PL
WARNING: the following files are missing in your kit:
        Makefile.PL
Please inform the author.
```

The change implemented here fixes this issue.